Rename 'decorator' module to avoid conflict with Python library
A 'decorator' library is already provided by Python. Having a
'decoratr' module of our own has caused import conflicts, e.g. obspy is
having trouble importing decorators.
