+++ mtpy/imaging/penetration_depth3d.py
@@ -21,7 +21,7 @@ import numpy as np
 
 import mtpy.core.mt as mt
 from mtpy.imaging.penetration import get_index, load_edi_files, Depth3D
-from mtpy.utils.decorator import deprecated
+from mtpy.utils.mtpy_decorator import deprecated
 from mtpy.utils.mtpylog import MtPyLog
 import logging
